+++ server/interfaces/Tree.ts
@@ -2,4 +2,5 @@ export default interface Tree {
   id: number;
   lat: number;
   lon: number;
+  time_created: string;
 }