+++ server/models/Tree.ts
@@ -27,20 +27,21 @@ function getByFilter(
  we just put the newest, verified tree
  */
 function getFeaturedTree(treeRepository: TreeRepository) {
-  return async function () {
-    const trees = await treeRepository.getByFilter(
-      {
-        approved: true,
-      },
-      { limit: 10, orderBy: { column: 'created_at', direction: 'desc' } },
-    );
-    // const trees: Array<Tree> = [];
-    // // eslint-disable-next-line no-restricted-syntax
-    // for (const id of [186737, 186735, 186736, 186734]) {
-    //   // eslint-disable-next-line no-await-in-loop
-    //   const tree = await treeRepository.getById(id);
-    //   trees.push(tree);
-    // }
+  return async () => {
+    // const trees = await treeRepository.getByFilter(
+    //  {
+    //    approved: true,
+    //  },
+    //  { limit: 10, orderBy: { column: 'time_created', direction: 'desc' } },
+    // );
+    //
+    const trees: Array<Tree> = [];
+    // eslint-disable-next-line no-restricted-syntax
+    for (const id of [186737, 186735, 186736, 186734]) {
+      // eslint-disable-next-line no-await-in-loop
+      const tree = await treeRepository.getById(id);
+      trees.push(tree);
+    }
     return trees;
   };
 }